proxy_platform/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.3
django-cors-headers==4.3.1
kombu==5.3.6
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
